dev_requirements.txt
	click==7.0
	setuptools_scm==3.2.0
	attrs==18.2.0
	bump2version==0.5.10
	gitpython==2.1.11
	nu-gitflow==1.0.2
	pylint==1.9.4
	six==1.12.0
	rope==0.14.0
	pytest>=4.6.5
	configparser
	pytest-xdist==1.29.0
